Safari-611.1.10/LayoutTests/http/tests/is-logged-in/available-in-secure-contexts.https.html

<!DOCTYPE html>
<html lang="en">
<head>
    <meta charset="UTF-8">
    <script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
    description("Tests that IsLoggedIn is available in secure contexts.");
    if (navigator.setLoggedIn !== undefined && typeof navigator.setLoggedIn === "function")
        testPassed("navigator.setLoggedIn is defined as a function on a page with protocol " + document.location.protocol);
    else
        testFailed("navigator.setLoggedIn is undefined or not of type function on a page with protocol " + document.location.protocol);
</script>
</body>
</html>
