LayoutTests/http/tests/is-logged-in/available-in-secure-contexts.https-expected.txt

Tests that IsLoggedIn is available in secure contexts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.setLoggedIn is defined as a function on a page with protocol https:
PASS successfullyParsed is true

TEST COMPLETE
